Update README.md
I am curious why you would suggest that. I have just spend useless time debugging why my path could not be found, because I blindly copied that command and thought it was just a 1:1 mapping.

If podlet does not support user systemd directories, it should also not support that in the container due to some strange mapping, should it?
If it does support that, then no reason to re-map it, it should work as usual, should not it?