Readme container doc: Also use z suffix for directories too
You suggest to disable SELinux completely for that feature. However the [`:Z` suffix](https://unix.stackexchange.com/questions/651198/podman-volume-mounts-when-to-use-the-z-or-z-suffix) can also be used for directories, so there is no need for that.

I've tested the provided command and it properly detected an existing file and warned from overwriting and could also write a new file. So it works.