feat(compose)!: rename `--pod` to `--kube`
`podlet compose --pod <POD>` is now `podlet compose --kube`.

The `name` field of a compose file is required when using `--kube`.
It is used as the name of the pod and in the filenames of the generated
`.kube` quadlet file and Kubernetes YAML file.

Frees `podlet compose --pod` for creating a `.pod` quadlet file from a
compose file in the future.

Signed-off-by: Paul Nettleton <k9@k9withabone.dev>