fix(compose): prepend container dependencies with pod name
When the `podlet compose --pod` option is used, the names of the
services from the Compose file are prepended with the pod name. If a
service had a dependency via the `depends_on` attribute, the name of the
dependency was not similarly prepended with the pod name when added to
the `[Unit]` section of the Quadlet file.

Changed `podlet::cli::compose::service_try_into_quadlet_file()` to
handle all service name changes when the `--pod` option is used.

Fixes: #114
Signed-off-by: Paul Nettleton <k9@k9withabone.dev>